backend/requirements.txt
Django>==3.2.3
django-filter>=23.*
django-colorfield>==0.8.0
djangorestframework>==3.14.0
djoser>==2.1.0
drf-extra-fields>==3.4.1
isort>=5.12.0
gunicorn>=20.1.0
webcolors>==1.11.1
psycopg2-binary>==2.9.6
Pillow>==9.0.0
pytest>==6.2.4
pytest-django>==4.4.0
pytest-pythonpath>==0.7.3
python-dotenv>=1.0.0